clair vele help cli
